+++ lib/supply_chain.dart
@@ -9,7 +9,7 @@ library;
 export 'src/doc.dart';
 export 'src/graph.dart';
 export 'src/insert.dart';
-export 'src/inserts.dart';
+export 'src/plugin_inserts.dart';
 export 'src/keys.dart';
 export 'src/node_blue_print.dart';
 export 'src/node.dart';